clean code, add dependencies, fix bugs.
