+++ pixels-common/src/main/java/io/pixelsdb/pixels/common/physical/impl/scheduler/SortMergeScheduler.java
@@ -53,6 +53,10 @@ public class SortMergeScheduler implements Scheduler
     @Override
     public CompletableFuture<Void> executeBatch(PhysicalReader reader, RequestBatch batch) throws IOException
     {
+        if (batch.size() <= 0)
+        {
+            return batch.completeAll();
+        }
         List<CompletableFuture<ByteBuffer>> futures = batch.getFutures();
         List<Request> requests = batch.getRequests();
         List<RequestFuture> requestFutures = new ArrayList<>(batch.size());
